🤖 refactor: Effect Phase 11 PR 4b — staged core layers S4–S8 + CoreWiringLive (imperative core body removed)
Peel the remainder of the core graph into Layer.effect adapters
(MemoryConsolidation · MCPConfig → MCPServerManager → Workspace → Task →
WorkspaceTurnManager) and replay the former body's setter/listener wiring,
in its original order, in CoreWiringLive. CoreLive = CoreWiringLive ▹ S8;
buildCoreTail and the transitional projection are gone, coreServices.ts is
types only.
